+++ shaders/test.wgsl
@@ -4,9 +4,9 @@ struct FrameData {
 };
 
 struct LED {
-    r: u32;
-    g: u32;
-    b: u32;
+    r: f32;
+    g: f32;
+    b: f32;
 };
 
 struct Result {
@@ -38,9 +38,9 @@ fn main([[builtin(global_invocation_id)]] global_id: vec3<u32>) {
     var end: u32 = min(index + 64u, arrayLength(&result.leds));
 
     loop {
-        result.leds[index].r = u32(positions.data[index].x);
-        result.leds[index].g = u32(positions.data[index].y);
-        result.leds[index].b = u32(positions.data[index].z);
+        result.leds[index].r = positions.data[index].x;
+        result.leds[index].g = positions.data[index].y;
+        result.leds[index].b = positions.data[index].z;
 
         index = index + 1u;
